requirements.txt
PySide6
markdown
markdown-it-py
pygments
gitpython
python-dotenv
